Removed Main2Activity
